Ack all events after receiving the entire set, to force complete retransmit on failure
Deferring the acknowledgement of incoming events until all events in a stream are received
and processed greatly simplifies handling of retries. While it would, in some ways, be
more efficient if we were able to resume transmission in a way that only required missed
events to be retransmitted, this would complicate matters in others. For one, we would
need to persist data as we received the events. We would also need to handle the scenario
where the retry is sent to a different instance of the server, in a setting where the
server is scaled out horizontally.

While there are certainly solutions we might investigate that allow us to support resuming on retry,
it is for now far simpler to require a complete retransmission of events.