Avoid repeated processing if the client retries
We now ack all events after the entire stream of events is received and
processed. This means that if the client retries, it can be expected
to retry from the start of the stream such that any new request
should always start with sequence number 1.

If, however, our response to the client is dropped, it is possible
for the client to begin its retry from some other position. In this
situation we would already have persisted the event stream. If this occurs
we should simply skip any processing and ack the events.