Fix operations count not showing correctly
The design of `buildqueuestateproxy` is to only show content with
instance names that the user is allowed to see, as specified by the
`instanceNameAuthorizer`. The pagination info was intentionally set to 0
to now disclose information about the part of the system that the user
should not have access to. This caused to the total operations count to
stop working.

This commit fixes the operations count by
fetching all operations from the scheduler and recreating the pagination
info based on the operations that the user is allowed to see.

fixes buildbarn/bb-portal#94