Incremental BES
Rewrites the BES handling to save each message as
it arrives instead of buffering them all in memory
and then processing them at once when the stream
ends. This makes it possible to show in-progress
invocations in the UI. It it also a nesscary step
towards the bb-portal backend being stateless.

NOTE: This is a breaking change that requires
requires a clean database to work. All existing
data will be lost.