Remove the build graph metrics visualization
The build graph metrics event describes the metrics of bazels internal
skyframe state graph. While it is information that is sent by bazel via
the build event stream it is unlikely to be useful to anyone who is not
debugging bazel itself. Since it is unlikely to be relevant for
developers and only adds to confusion we have opted to delete the
visualization.