Merge entity `FilesMetric` into `ArtifactMetrics`
The automatic cascade delete of `FilesMetric` when
deleting an `ArtifactMetrics` does not work
properly. Instead of ent creating 4 back edges
from `FilesMetric` to `ArtifactMetrics`, it
creates only one. This means that when deleting
an `ArtifactMetrics`, only one `FilesMetric`
is deleted, and the rest remain in the database
as dangling references.

Instead of trying to fix this, we remove
`FilesMetric` and move its fields into
`ArtifactMetrics`.