Rework graphql authorization
The old authorization system was based on ent interceptors, which
modified the responses from the database layer to filter out
unauthorized data. This approach had several issues:

- Performance.
- Now allowing for database queries that return only integers, such as
  counts.
- No easy way to bypass authorization for internal services.

This commit replaces the old interceptor-based authorization system with
a new privacy-based system. The new system uses ent privacy rules to
enforce authorization by modifying the queries sent to the database.
When you make a query, the `DbAuthService` will fech all `instanceNames`
in the database (and cache them for future queries) and run them through
the `instanceNameAuthorizer` to determine which `instanceNames` the user
has access to. The privacy rules will then modify the queries to only
return data for the allowed `instanceNames`.

This solves the issues with the old system, but introduces some new
ones:

- There is a limit on how many `instanceNames` that a user can have
  access to, due to the limit on how many parameters a SQL query can
  have. By default, both SQLite and Postgres can handle at leaset 32000
  parameters, so this should be sufficient for most use cases.
- The performance of queries gets worse the more `instanceNames` a user
  has access to. In our tests, 32000 `instanceNames` results in queries
  taking an additional ~100ms. If this becomes an issue there are
  several possible optimizations, such as caching the allowed
  `instanceNames` per user, either in memory or in the database (so that
  we don't have to send approved `instanceNames` as query parameters).