Add sqlc support
Adds a generic database client that allows us to use both sqlc and
ent with the same underlying database connection and transactions.

To facilitate this the `*ent.Client` type has been mostly replaced by
our `database.Client` interface which offers the ability to ask for
either a sqlc client or an ent client as well as start a transaction
spanning both tech stacks.

The big advantage of this is the ability for us to add strongly typed
raw sql for more advanced use cases. We can then use sqlc tooling to
generate the corresponding boiler plate to serialize and deserialize the
sql.

As our code is becoming more sql native maintaining both sqlite and
postgres support is becoming untenable. This commit removes the Sqlite
support in favor of postgres native solutions.